(add) Add S3 to updateFile API
