+++ app.js
@@ -3,7 +3,7 @@ import path from 'path';
 import cookieParser from 'cookie-parser';
 import morgan from 'morgan';
 
-import filesRouter from './src/routes/filesRoute.js';
+import fileRouter from './src/routes/fileRoute.js';
 
 const app = express();
 const __dirname = path.resolve();
@@ -14,6 +14,6 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/files', filesRouter);
+app.use('/file', fileRouter);
 
 export default app;
