+++ package.json
@@ -14,6 +14,8 @@
     "dotenv": "^9.0.2",
     "express": "~4.16.1",
     "morgan": "~1.9.1",
+    "multer": "^1.4.2",
+    "multer-s3": "^2.9.0",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
